% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.summary.R
\name{benchmark.summary}
\alias{benchmark.summary}
\title{benchmark.summary}
\usage{
benchmark.summary(benchmark_dir)
}
\arguments{
\item{benchmark_dir}{String indicating the full path to the \code{benchmarks} folder in the output directory of the snakemake pipeline.}
}
\value{
A data frame with the steps (rules) and the respective time and resources spent.
}
\description{
Prints a summary of resources and time used to complete the pipeline. It reads the files in the \code{benchmarks} folder.
}
