% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.BINDetect.volcano.R
\name{plot.BINDetect.volcano}
\alias{plot.BINDetect.volcano}
\title{plot.BINDetect.volcano}
\usage{
\method{plot}{BINDetect.volcano}(
  results,
  motif.pattern = "_(MOUSE|HUMAN)[.]H[0-9]*MO",
  right.color = "indianred",
  left.color = "steelblue",
  ns.color = "gray30",
  points.transparency = 0.5,
  labels.min.segment.length = 0,
  add.signif.lables = TRUE,
  labels.repel.force = 10,
  labels.max.overlaps = 100,
  extra.labels = "none"
)
}
\arguments{
\item{results}{Path to the TOBIAS BINDetect results table, e.g. \code{bindetect_results.txt}, or a corresponding data.frame.}

\item{motif.pattern}{String of the regular expression indicating the pattern to remove from the motif ID. Default: \code{"_(MOUSE|HUMAN)[.]H[0-9]*MO"}.}

\item{right.color}{String indicating the color to use for the differential motifs in the right side (positive difference) of the plot. Default: \code{"indianred"}.}

\item{left.color}{String indicating the color to use for the differential motifs in the left side (negative difference) of the plot. Default: \code{"steelblue"}.}

\item{ns.color}{String indicating the color to use for the non-significant motifs of the plot. Default: \code{"gray30"}.}

\item{points.transparency}{Numeric value between 0-1 indicating the transparency (alpha) of the dots in the volcano. Default: \code{0.5}.}

\item{labels.min.segment.length}{Numeric value indicating the minimum length of the line connecting the motif labels and the corresponding dot (passed to \code{ggrepel}). Default: \code{0}, segment always present.}

\item{add.signif.lables}{Logical value indicating whether differential motifs should be labelled. Default: \code{TRUE}.}

\item{labels.repel.force}{Numeric value indicated the force of repulsion between overlapping text labels (passed to \code{ggrepel}). Default: \code{10}.}

\item{labels.max.overlaps}{Numeric value indicating of maximum labels overlapping allowed (passed to \code{ggrepel}). Default: \code{100}.}

\item{extra.labels}{Vector of other motifs to display, retrospectively by their significance. Default: \code{"none"}.}
}
\value{
A ggplot object.
}
\description{
Function to replot the volcano plot of differential motifs (footprints) as generated by BINDetect.
}
