% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.correlation.R
\name{plot.correlation}
\alias{plot.correlation}
\title{plot.correlation}
\usage{
\method{plot}{correlation}(
  correlation.matrix,
  palette = viridis::mako(100, direction = -1),
  correlation.scale.limits = c(0, 1),
  exclude.diagonal = FALSE,
  dendrogram.position = "left",
  dendrogram.color = "black",
  dendrogram.linewidth = 0.5,
  display.values = TRUE,
  values.color = "white",
  values.decimals = 2,
  values.font.size = 2,
  values.transparency = 1,
  plot.title = NULL,
  plot.subtitle = NULL,
  clustering.method = "complete",
  return.cluster.object = FALSE
)
}
\arguments{
\item{correlation.matrix}{Path or data.frame of the correlation matrix generated by snakeATAC.}

\item{palette}{Vector of colors corresponding to the palette to use for the heatmap color scale. Default: \code{viridis::mako(100, direction = -1)}.}

\item{correlation.scale.limits}{Two-elements vector to indicate lower and higher limits, respectively, to apply to the correlation coefficient color scale. Default: \code{c(0,1)}.}

\item{dendrogram.position}{String indicating the position of the dendrogram. One among: "top", "bottom", "left", "right". Default: \code{"left"}.}

\item{dendrogram.color}{String indicating the color of the dendrogram lines. Default: \code{"black"}.}

\item{dendrogram.linewidth}{Numeric value indicating the line.width of the dendrogram. Default: \code{"black"}.}

\item{display.values}{Logical value indicating whether the correlation coefficient should be displayed for each cell. Default: \code{TRUE}.}

\item{values.color}{String indicating the color to use for the correlation coefficient labels. Default: \code{"white"}.}

\item{values.decimals}{Numeric value indicating the number of decimals at which round the correlation coefficient labels. Default: \code{2}.}

\item{values.font.size}{Numeric value indicating the font size of the correlation coefficient labels. Default: \code{2}.}

\item{values.transparency}{Numeric value between 0-1 indicating the transparency (alpha) of the correlation coefficient labels. Default: \code{1}, full color.}

\item{plot.title}{String indicating the main title of the plot. Default: \code{NULL}.}

\item{plot.subtitle}{String indicating the subtitle of the plot. Default: \code{NULL}.}

\item{clustering.method}{String indicating the clustering method to use. The value should be (an unambiguous abbreviation of) one among: 'ward.D', 'ward.D2', 'single', 'complete', 'average' (= UPGMA), 'mcquitty' (= WPGMA), 'median' (= WPGMC) or 'centroid' (= UPGMC).}

\item{return.cluster.object}{Logical value indicating whether the clustering and distance object should be returned along with the plot. Default: \code{FALSE}, only the plot is returned.}
}
\value{
Either a ggplot object, or a list with ggplot object, distance and hclust objects
}
\description{
plot.correlation
}
