% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.footprint.R
\name{plot.footprint}
\alias{plot.footprint}
\title{plot.footprint}
\usage{
\method{plot}{footprint}(
  matrix.gz,
  factor = NULL,
  colors = NULL,
  normalize.zero.one = FALSE,
  white.list = NULL
)
}
\arguments{
\item{matrix.gz}{Path to a motif matrix generated by snakeATAC. Usually it can be find at \code{./05b_Differential_TF_binding_TOBIAS/D_density_profiles_merged_BAMs/matrices/MOTIF.ID_single.base.scores_per.region.gz}.}

\item{factor}{String indicating the name of the Transcription factor, used as title of the plot. Default: \code{NULL}.}

\item{colors}{Vector of colors, or named color vector with the colors of each group, to use for the line plots. Default: \code{NULL}.}

\item{normalize.zero.one}{Logical value indicating whether profiles should be normalized between 0 and 1. Default: \code{FALSE}, raw data.}

\item{white.list}{A path to a bed file, or a bed file in data.frame or GRanges format. This region list will be used to filter the full table and only the average over these regions will be plotted. Default: \code{NULL}, no filtering.}
}
\value{
A ggplot object.
}
\description{
Replotting of footprints starting from \href{https://deeptools.readthedocs.io/en/develop/}{deeptools} matrices. It allows for the filter over a specific region set (white list).
}
