% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.PCA.R
\name{plot.PCA}
\alias{plot.PCA}
\title{plot.PCA}
\usage{
\method{plot}{PCA}(
  matrix,
  excluded.samples = NULL,
  scale = TRUE,
  center = TRUE,
  PCs = c(1, 2),
  point.size = 3,
  point.alpha = 0.75,
  color.list = NULL
)
}
\arguments{
\item{matrix}{Matrix where each column is a sample and each row a genomic bin/region. A multiBigwigSummary file can be read using \link{read.multibigwigsummary}.}

\item{excluded.samples}{String vector indicating the list of sample IDs to exclude from the plot. Default: \code{NULL}.}

\item{scale}{Logical value to indicate whether the data should be scaled. Default: \code{TRUE}.}

\item{center}{Logical value to indicate whether the data should be centered. Default: \code{TRUE}.}

\item{PCs}{Numeric vector of length 2, indicating which principal components (PCs) should be plotted. Default: \code{c(1,2)}.}

\item{point.size}{Numeric value indicating the dot size. Default: \code{3}.}

\item{point.alpha}{Numeric value between 0-1 indicating the dot transparency (1 full, 0 invisible). Default: \code{0.75}.}

\item{color.list}{(Named) String vector indicating any R-supported color to assign to each sample. Default: \code{NULL} (automatic).}

\item{shape.list}{Vector indicating any R-supported shape to assign to each sample. Default: \code{NULL} (automatic).}
}
\value{
A ggplot2 object.
}
\description{
Computes and plots PCA starting from a multiBigwigSummary file.
}
