% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.multibigwigsummary.R
\name{read.multibigwigsummary}
\alias{read.multibigwigsummary}
\title{read.multibigwigsummary}
\usage{
read.multibigwigsummary(multibigwigsummary.file)
}
\arguments{
\item{multibigwigsummary.file}{Path to a \code{".npz"} file generated by \code{multiBigwigSummary}.}
}
\value{
A matrix where each column is a sample and each row a genomic bin/region.
}
\description{
Imports ".npz" files generated by \code{multiBigwigSummary} (\code{deepTools}).
}
